* master: (84 commits)
  Introduce ASSERT_OR_RETURN, a macro to assert on an expression, and assert() and then return if it fails,
  The commit to fix the unsigned int underflow (ticket 404) had as unfortunate side effect that building was possible even when out of power. This fixes it.
  Make BinaryProtocol more flexible to allow it to work with 2.0+ as well as the 2.2 lobby protocol (depending on what parameter was used during construction)
  Move the Game, Protocol classes to files of their own (to allow them to be reused by the IRC bot, instead of duplicating code)
  Be extra careful with the buildpoints, as the signed <-> unsigned stuff might have caused problems. This should fix ticket #404.
  Move the IRC bot to the same directory as the lobby server
  Whatever way we leave the bot's mainloop, make sure to terminate the lobby checking thread
  Don't use Python 2.6 specific syntax (r7123)
  * Catch and handle all socket errors, not just timeouts
  excise the codebase from some C99 format specifiers.
  Fix #400:
  * Notify all served IRC channels when shutting down
  * Rename the "help" command to "about"
  * Move construction of BotCommands to its own class (__init__)
  * Pass the IRC write function as a lambda to BotCommands' functions
  Properly store the gameFormat struct description in the class
  * Add channel operator support
  * Send replies only to the channel from which the request came
  Don't call the IRC command handling function in the try block (we don't want to protect against errors inside the IRC command function)
  Add multi-channel support to wzlobbybot, one list of channels in which to serve requests, another just to be present in.
  ...

+++ src/multiplay.c
@@ -536,7 +536,6 @@ BOOL myResponsibility(UDWORD player)
 //returns true if 'player' is responsible for 'playerinquestion'
 BOOL responsibleFor(UDWORD player, UDWORD playerinquestion)
 {
-
 	if(whosResponsible(playerinquestion) == player)
 	{
 		return true;
@@ -1894,7 +1893,7 @@ const char* getPlayerColourName(unsigned int player)
 		N_("Cyan")
 	};
 
-	ASSERT(player < ARRAY_SIZE(playerColors), "player number (%d) exceeds maximum (%zu)", player, ARRAY_SIZE(playerColors));
+	ASSERT(player < ARRAY_SIZE(playerColors), "player number (%d) exceeds maximum (%lu)", player, (unsigned long) ARRAY_SIZE(playerColors));
 
 	if (player >= ARRAY_SIZE(playerColors))
 	{
